+++ libs/data/farmer_content/data/content/1_what_you_do.ts
@@ -1,0 +1,2 @@
+const content: IFarmerContent = {};
+export default content;
